Add FBR Annexure-A purchase import (preview + commit)
Operator uploads the monthly FBR Annexure-A xls; preview decides each
row (will-import / already-claimed / unregistered-seller / no-hs-code /
etc.); commit auto-creates Suppliers, Purchase Bills, PurchaseItems,
ItemTypes, and StockMovements per-invoice transactionally. Bad invoices
roll back without poisoning the batch.

Schema: PurchaseItem.Quantity widened int → decimal(18,4) so fractional
UOMs survive; +PurchaseItem.ExtraTax / StWithheldAtSource for FBR
line-tax fields; +PurchaseBill.Source tagging the row's lineage;
+ItemType.IsHsCodePartial flagging 4-digit-only HS codes that PRAL
rejects for sales until upgraded to a full PCT.

+++ DTOs/PurchaseBillDto.cs
@@ -39,7 +39,7 @@ namespace MyApp.Api.DTOs
         public int? ItemTypeId { get; set; }
         public string ItemTypeName { get; set; } = "";
         public string Description { get; set; } = "";
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public string UOM { get; set; } = "";
         public decimal UnitPrice { get; set; }
         public decimal LineTotal { get; set; }
@@ -79,7 +79,7 @@ namespace MyApp.Api.DTOs
     {
         public int? ItemTypeId { get; set; }
         public string Description { get; set; } = "";
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public string? UOM { get; set; }
         public decimal UnitPrice { get; set; }
         public string? HSCode { get; set; }
@@ -121,7 +121,7 @@ namespace MyApp.Api.DTOs
         public int Id { get; set; }
         public int? ItemTypeId { get; set; }
         public string Description { get; set; } = "";
-        public int Quantity { get; set; }
+        public decimal Quantity { get; set; }
         public string UOM { get; set; } = "";
         public decimal UnitPrice { get; set; }
         public string? HSCode { get; set; }
